requirements.txt
beanie==1.29.0
fastapi==0.115.12
motor==3.7.0
passlib==1.7.4
pydantic==2.11.2
python_jose==3.4.0
uvicorn==0.34.0
